requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
django-environ==0.11.2
django-allauth
requests
jwt
